requirements.txt
flask>=3.0
pypdf>=4.0
requests>=2.31
google-genai>=1.0
python-dotenv
gunicorn>=22.0